% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimatedataQAQC.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Not a number search, for dataframes}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{The dataframe itself}
}
\description{
Function to find nan values in a dataframe
}
\examples{
a <- c(NaN, 1)
b<- c(2, 3)
df<-cbind.data.frame(a, b)
is.nan.data.frame(df) #returns:
         a     b
[1,]  TRUE  FALSE
[2,] FALSE  FALSE
}
