% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimatedataQAQC.R
\name{sumna}
\alias{sumna}
\title{Sum not a number values}
\usage{
sumna(x)
}
\arguments{
\item{x}{A list of values}
}
\description{
Sums values, returns NA if all values are NA. Builds on the sum() function, which returns 0 if all values are NA.
}
\examples{
a<-c(NA, NA, NA, NA)
b<-c(1, NA, NA, 2)
sumna(a) #returns:
[1] NA
sumna(b) #returns:
[1] 3
}
